Implement new rate limit API
