+++ buildSrc/src/main/kotlin/mihon.code.lint.gradle.kts
@@ -27,6 +27,7 @@ spotless {
                 "ktlint_function_naming_ignore_when_annotated_with" to "Composable",
                 "ktlint_standard_class-signature" to "disabled",
                 "ktlint_standard_discouraged-comment-location" to "disabled",
+                "ktlint_standard_filename" to "disabled",
                 "ktlint_standard_function-expression-body" to "disabled",
                 "ktlint_standard_function-signature" to "disabled",
             ))
